renamed setup and teardown file
