+++ test/Attributes/AttributeKeys.test.ts
@@ -1,6 +1,6 @@
 import { repoId } from '../testHelper.js';
 import { ODataValueContextOfListOfAttribute, IRepositoryApiClient } from '../../src/index.js';
-import { _RepositoryApiClient } from '../setup.js';
+import { _RepositoryApiClient } from '../createSession.js';
 
 
 describe('Attribute Key Integration Tests', () => {
