+++ test/Entries/RemoveEntries.test.ts
@@ -1,7 +1,7 @@
 import { repoId } from '../testHelper.js';
 import { DeleteEntryWithAuditReason, Entry, PutTemplateRequest } from '../../src/index.js';
 import { allFalse, CreateEntry } from '../BaseTest.js';
-import { _RepositoryApiClient } from '../setup.js';
+import { _RepositoryApiClient } from '../createSession.js';
 
 describe('Remove Entries Integration Tests', () => {
   let entry = new Entry();
