+++ test/FieldDefinitions/GetFieldDefinitions.test.ts
@@ -1,6 +1,6 @@
 import { repoId } from '../testHelper.js';
 import { ODataValueContextOfIListOfWFieldInfo, IRepositoryApiClient, WFieldInfo } from '../../src/index.js';
-import { _RepositoryApiClient } from '../setup.js';
+import { _RepositoryApiClient } from '../createSession.js';
 
 describe('Field Definitions Integration Tests', () => {
   test('Get Field Definitions', async () => {
