+++ test/Searches/CloseSearchTests.test.ts
@@ -1,6 +1,6 @@
 import { repoId } from '../testHelper.js';
 import { AdvancedSearchRequest, ODataValueContextOfIListOfEntry } from '../../src/index.js';
-import { _RepositoryApiClient } from '../setup.js';
+import { _RepositoryApiClient } from '../createSession.js';
 
 describe('Search Integration Tests', () => {
   test('Close Search Operations', async () => {
